fix: resolve 45 analyzability violations — split files, reduce nesting (#117)
Python:
- Extract _incremental.py from runner.py (481→288 lines)
- Extract _consolidated.py from subagents/runner.py (361→284 lines)
- Extract _helpers.py from menubar.py (347→287 lines)
- Reduce pool.py (368→297), subprocess.py (302→298), routes.py (306→299)
- Split long functions into focused helpers (≤50 lines each)
- Flatten nesting in verify.py, pool.py, project_resolver.py, validation.py
- Remove unused IS_WIN32 import, narrow bare except blocks
- Add missing docstrings and blank lines

UI:
- Extract MainContent from App.jsx (228→<50 line function)
- Extract sub-components: DimensionCardsGrid, AboutSection, ViolationRow, etc.
- Replace IIFE with variable in RunOverviewPanel
- Use array.some() for boolean chains in EvaluationStatus

Tests:
- Split test_evidence_parser.py (340→269) and test_subagent_pool.py (346→195)
- Move late imports to top of test_file_priority.py
