+++ src/quodeq/core/types/_mapper_findings.py
@@ -18,6 +18,7 @@ from ._mapper_helpers import (
 
 
 def parse_req_ref(raw: dict[str, object]) -> ReqRef:
+    """Parse a raw dict into a ReqRef (requirement reference) instance."""
     return ReqRef(
         label=_str(raw, "label"),
         url=_str(raw, "url"),
