+++ src/quodeq/ui/src/features/dashboard/components/ViolationsByPrincipleTable.jsx
@@ -4,6 +4,33 @@
 
 import { memo, useMemo, useState } from 'react';
 
+function PrincipleRow({ p, idx, onPrincipleClick }) {
+  return (
+    <li
+      key={p.principle || idx}
+      className={`offending-file-row${onPrincipleClick ? ' offending-file-row--clickable' : ''}`}
+      role={onPrincipleClick ? 'button' : undefined}
+      tabIndex={onPrincipleClick ? 0 : undefined}
+      onClick={onPrincipleClick ? () => onPrincipleClick(p) : undefined}
+      onKeyDown={onPrincipleClick ? (e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); onPrincipleClick(p); } } : undefined}
+    >
+      <div className="offending-file-info">
+        <span className="offending-file-path">{p.principle}</span>
+        {p.dimensionsStr && (
+          <span className="offending-file-dims">{p.dimensionsStr}</span>
+        )}
+      </div>
+      <strong className="offending-file-total">{p.total}</strong>
+      <span className="offending-file-tags">
+        {p.critical > 0 && <span className="severity-tag critical">{p.critical} critical</span>}
+        {p.major > 0 && <span className="severity-tag major">{p.major} major</span>}
+        {p.minor > 0 && <span className="severity-tag minor">{p.minor} minor</span>}
+      </span>
+      {onPrincipleClick && <span className="offending-file-chevron">›</span>}
+    </li>
+  );
+}
+
 const ViolationsByPrincipleTable = memo(function ViolationsByPrincipleTable({ violations, onPrincipleClick, pageSize = 20 }) {
   const [showAll, setShowAll] = useState(false);
 
@@ -44,28 +71,7 @@ const ViolationsByPrincipleTable = memo(function ViolationsByPrincipleTable({ vi
     <>
       <ul className="offending-file-list">
         {displayItems.map((p, idx) => (
-          <li
-            key={p.principle || idx}
-            className={`offending-file-row${onPrincipleClick ? ' offending-file-row--clickable' : ''}`}
-            role={onPrincipleClick ? 'button' : undefined}
-            tabIndex={onPrincipleClick ? 0 : undefined}
-            onClick={onPrincipleClick ? () => onPrincipleClick(p) : undefined}
-            onKeyDown={onPrincipleClick ? (e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); onPrincipleClick(p); } } : undefined}
-          >
-            <div className="offending-file-info">
-              <span className="offending-file-path">{p.principle}</span>
-              {p.dimensionsStr && (
-                <span className="offending-file-dims">{p.dimensionsStr}</span>
-              )}
-            </div>
-            <strong className="offending-file-total">{p.total}</strong>
-            <span className="offending-file-tags">
-              {p.critical > 0 && <span className="severity-tag critical">{p.critical} critical</span>}
-              {p.major > 0 && <span className="severity-tag major">{p.major} major</span>}
-              {p.minor > 0 && <span className="severity-tag minor">{p.minor} minor</span>}
-            </span>
-            {onPrincipleClick && <span className="offending-file-chevron">›</span>}
-          </li>
+          <PrincipleRow key={p.principle || idx} p={p} idx={idx} onPrincipleClick={onPrincipleClick} />
         ))}
       </ul>
       {hasMore && (
