+++ tests/engine/test_subagent_pool_advanced.py
@@ -1,0 +1,181 @@
+"""Tests for SubagentPool — scout-then-scale and pool budget."""
+from __future__ import annotations
+
+import json
+from pathlib import Path
+from unittest.mock import patch
+
+import pytest
+
+from quodeq.analysis.subprocess import AnalysisConfig, AnalysisError
+from quodeq.engine.file_queue import FileQueue
+from quodeq.engine.subagent_pool import PoolPaths, SubagentPool, SubagentResult
+
+
+def _fake_run_analysis(work_dir, prompt, stream_file, config):
+    """Mock run_analysis that writes some findings and drains the queue."""
+    stream_file.parent.mkdir(parents=True, exist_ok=True)
+    stream_file.write_text("")
+    if config.queue_path:
+        queue = FileQueue(config.queue_path)
+        queue.take(queue.remaining(), agent_id=config.agent_id)
+    if config.jsonl_file:
+        agent_id = config.agent_id or "unknown"
+        with open(config.jsonl_file, "a") as f:
+            f.write(json.dumps({
+                "schema_version": 1,
+                "p": "Modularity", "t": "violation", "d": "maintainability",
+                "w": f"Found by {agent_id}", "file": f"{agent_id}.py", "line": 1,
+            }) + "\n")
+
+
+class TestComputeScaleUp:
+    def _make_pool(self, n_agents, tmp_path, max_files=30):
+        queue_path = tmp_path / "queue.json"
+        FileQueue(queue_path, ["f.py"])
+        return SubagentPool(
+            n_agents=n_agents,
+            paths=PoolPaths(work_dir=tmp_path, evidence_dir=tmp_path, queue_path=queue_path),
+            prompt="test",
+            dimension="security",
+            config=AnalysisConfig(max_files_per_agent=max_files),
+        )
+
+    def test_no_remaining_returns_zero(self, tmp_path):
+        pool = self._make_pool(5, tmp_path)
+        assert pool._compute_scale_up(0) == 0
+
+    def test_remaining_within_one_batch(self, tmp_path):
+        pool = self._make_pool(5, tmp_path)
+        assert pool._compute_scale_up(25) == 0
+
+    def test_remaining_needs_two_agents(self, tmp_path):
+        pool = self._make_pool(5, tmp_path)
+        assert pool._compute_scale_up(50) == 2
+
+    def test_remaining_capped_by_max_agents(self, tmp_path):
+        pool = self._make_pool(3, tmp_path)
+        assert pool._compute_scale_up(200) == 2
+
+    def test_max_agents_1_never_scales(self, tmp_path):
+        pool = self._make_pool(1, tmp_path)
+        assert pool._compute_scale_up(500) == 0
+
+
+class TestMultiDimensionPool:
+    def test_single_dimension_backward_compat(self, tmp_path):
+        """Single dimension string still works as before."""
+        queue_path = tmp_path / "queue.json"
+        FileQueue(queue_path, ["a.py"])
+        pool = SubagentPool(
+            n_agents=1,
+            paths=PoolPaths(work_dir=tmp_path, evidence_dir=tmp_path, queue_path=queue_path),
+            prompt="test",
+            dimension="security",
+        )
+        assert pool._dimension == "security"
+        assert pool._dimension_key == "security"
+
+    def test_multi_dimension_list(self, tmp_path):
+        """List of dimensions uses 'consolidated' key for file naming."""
+        queue_path = tmp_path / "queue.json"
+        FileQueue(queue_path, ["a.py"])
+        pool = SubagentPool(
+            n_agents=1,
+            paths=PoolPaths(work_dir=tmp_path, evidence_dir=tmp_path, queue_path=queue_path),
+            prompt="test",
+            dimension=["security", "maintainability"],
+        )
+        assert pool._dimension == "security,maintainability"
+        assert pool._dimension_key == "consolidated"
+        assert pool._dimensions == ["security", "maintainability"]
+
+    def test_multi_dimension_jsonl_path(self, tmp_path):
+        queue_path = tmp_path / "queue.json"
+        FileQueue(queue_path, ["a.py"])
+        pool = SubagentPool(
+            n_agents=1,
+            paths=PoolPaths(work_dir=tmp_path, evidence_dir=tmp_path, queue_path=queue_path),
+            prompt="test",
+            dimension=["security", "maintainability"],
+        )
+        assert "consolidated_evidence.jsonl" in str(pool._shared_jsonl_path())
+
+
+class TestScoutThenScale:
+    def test_small_queue_uses_one_agent(self, tmp_path):
+        """20 files with max_agents=5 -> only 1 agent should run (scout handles all)."""
+        queue_path = tmp_path / "queue.json"
+        FileQueue(queue_path, [f"src/f{i}.py" for i in range(20)])
+
+        pool = SubagentPool(
+            n_agents=5,
+            paths=PoolPaths(work_dir=tmp_path, evidence_dir=tmp_path, queue_path=queue_path),
+            prompt="analyse",
+            dimension="security",
+            config=AnalysisConfig(max_files_per_agent=30),
+        )
+
+        with patch("quodeq.analysis.subagents.pool.run_analysis", _fake_run_analysis):
+            results = pool.run()
+
+        assert len(results) == 1
+        assert results[0].agent_id == "agent-0"
+
+    def test_large_queue_scales_up(self, tmp_path):
+        """200 files with max_agents=5 -> scout + overflow agents."""
+        queue_path = tmp_path / "queue.json"
+        FileQueue(queue_path, [f"src/f{i}.py" for i in range(200)], max_files_per_agent=30)
+
+        pool = SubagentPool(
+            n_agents=5,
+            paths=PoolPaths(work_dir=tmp_path, evidence_dir=tmp_path, queue_path=queue_path),
+            prompt="analyse",
+            dimension="security",
+            config=AnalysisConfig(max_files_per_agent=30),
+        )
+
+        with patch("quodeq.analysis.subagents.pool.run_analysis", _fake_run_analysis):
+            results = pool.run()
+
+        assert len(results) > 1
+        assert results[0].agent_id == "agent-0"
+
+    def test_scout_first_false_launches_all_agents(self, tmp_path):
+        """When scout_first=False, all agents launch immediately."""
+        queue_path = tmp_path / "queue.json"
+        FileQueue(queue_path, [f"src/f{i}.py" for i in range(200)], max_files_per_agent=30)
+
+        pool = SubagentPool(
+            n_agents=3,
+            paths=PoolPaths(work_dir=tmp_path, evidence_dir=tmp_path, queue_path=queue_path),
+            prompt="verify",
+            dimension="security",
+            config=AnalysisConfig(max_files_per_agent=30),
+            scout_first=False,
+        )
+
+        with patch("quodeq.analysis.subagents.pool.run_analysis", _fake_run_analysis):
+            results = pool.run()
+
+        # All 3 agents should have run (not just scout + scale-up)
+        agent_ids = {r.agent_id for r in results}
+        assert len(agent_ids) >= 3
+
+
+class TestPoolBudget:
+    def test_pool_uses_pool_budget_not_max_duration(self, tmp_path):
+        """Pool time limit should use pool_budget, not max_duration."""
+        queue_path = tmp_path / "queue.json"
+        FileQueue(queue_path, [f"f{i}.py" for i in range(5)])
+
+        ac = AnalysisConfig(pool_budget=60, max_duration=1800, max_files_per_agent=30)
+        pool = SubagentPool(
+            n_agents=1,
+            paths=PoolPaths(work_dir=tmp_path, evidence_dir=tmp_path, queue_path=queue_path),
+            prompt="test",
+            dimension="security",
+            config=ac,
+        )
+        assert pool._base_config.pool_budget == 60
+        assert pool._base_config.max_duration == 1800
