+++ tools/enrich_standards.py
@@ -92,19 +92,38 @@ def _get_highest_id(reqs: list[dict], prefix: str) -> int:
     return highest
 
 
+def _enrich_sub_characteristic(
+    sc: dict, dim_mapping: dict, prefix: str, existing: set[int], dry_run: bool,
+) -> int:
+    """Add new CWEs to a single sub-characteristic. Returns count added."""
+    sc_name = sc["name"]
+    if sc_name not in dim_mapping:
+        return 0
+    highest = _get_highest_id(sc["requirements"], prefix)
+    added = 0
+    for req_text, cwe_ids in dim_mapping[sc_name]:
+        new_cwes = [c for c in cwe_ids if c not in existing]
+        if not new_cwes:
+            continue
+        highest += 1
+        new_req = {"id": f"{prefix}-{highest}", "text": req_text, "cwe": sorted(new_cwes)}
+        sc["requirements"].append(new_req)
+        existing.update(new_cwes)
+        added += len(new_cwes)
+        if dry_run:
+            print(f"  {new_req['id']}: {len(new_cwes)} CWEs \u2192 {sc_name}")
+    return added
+
+
 def enrich_dimension(
-    dimension: str,
-    dry_run: bool = True,
-    *,
-    mapping: dict | None = None,
-    prefix_map: dict | None = None,
+    dimension: str, dry_run: bool = True,
+    *, mapping: dict | None = None, prefix_map: dict | None = None,
 ) -> int:
     """Add missing CWEs to one ISO 25010 dimension file. Returns count added."""
     if mapping is None:
         mapping = _load_mapping()
     if dimension not in mapping:
         return 0
-
     filepath = STANDARDS_DIR / f"{dimension}.json"
     if not filepath.exists():
         print(f"  SKIP {dimension}: file not found")
@@ -115,39 +134,13 @@ def enrich_dimension(
     if prefix_map is None:
         prefix_map = _load_prefix_map()
     prefixes = prefix_map.get(dimension, {})
-    total_added = 0
-
-    for sc in data["sub_characteristics"]:
-        sc_name = sc["name"]
-        if sc_name not in mapping[dimension]:
-            continue
-
-        prefix = prefixes.get(sc_name, _UNKNOWN_PREFIX)
-        highest = _get_highest_id(sc["requirements"], prefix)
-
-        for req_text, cwe_ids in mapping[dimension][sc_name]:
-            # Filter out CWEs that already exist
-            new_cwes = [c for c in cwe_ids if c not in existing]
-            if not new_cwes:
-                continue
-
-            highest += 1
-            new_req = {
-                "id": f"{prefix}-{highest}",
-                "text": req_text,
-                "cwe": sorted(new_cwes),
-            }
-            sc["requirements"].append(new_req)
-            existing.update(new_cwes)
-            total_added += len(new_cwes)
-
-            if dry_run:
-                print(f"  {new_req['id']}: {len(new_cwes)} CWEs \u2192 {sc_name}")
-
+    total_added = sum(
+        _enrich_sub_characteristic(sc, mapping[dimension], prefixes.get(sc["name"], _UNKNOWN_PREFIX), existing, dry_run)
+        for sc in data["sub_characteristics"]
+    )
     if not dry_run and total_added > 0:
         filepath.write_text(json.dumps(data, indent=2, ensure_ascii=False) + "\n", encoding=_TEXT_ENCODING)
         print(f"  Wrote {filepath.name} (+{total_added} CWEs)")
-
     return total_added
